Added Background Smart Checking
Let's say you have a TitaniumBackup schedule coming up and your phone's screen is off. Normally this would execute suspend engine and leave it as that until you turn the screen back on. This will slow down background tasks performed by performance apps such as TitaniumBackup and will drag battery life down as it'll take longer to process. I have found an excellent solution to this issue. When TitaniumBackup starts, it creates a high priority notification, Helix Engine's Accessibility Service can now listen to notification state changes. This means that if TiB is started, it'll detect that it's started a background task. When it is detected, the profile that this app is assigned in is executed, bypassing suspend engine if your screen is off. This ensures that the background task performed by a performance app will get the job done as fast as possible. Once it has finished, the engine will detect that the notification has gone away, assuming that the background task has finished, it will execute suspend engine again (if your screen is still off). This solution will make this engine EVEN MORE efficient than it already is.