Added UI to Helix Engine
Also with this commit, ALL DEVICES are supported in one app, which means all the scripts are added into the assets folder. The app will load a set of scripts depending on the chipset the user chooses in the app UI.

Again, this is a debug build which enables logcat logging and apk modification. Refer back to previous commit if you forgot how to try out your new scripts in this apk.